Make it an SPA with a singleton session.
We actually want to support multiple users, of
course, and this sets the stage a little bit
to have some kind of session cookies.

It also makes it less annoying to test, since
we don't have constant redirects.  (Once we
figure out the oauth piece, it should just be
a black box, mostly.)