events: Implement event queue, message fetching and live rerendering.
This is one big commit that adds structurally sound foundation for a
realtime events implementation. As a test, we can now send someone a
PM and see it reflected live in the UI.

To do this, the proxy server registers one event queue for each client
and parcels any data updates (for now only messages) to be sent to the
client via their websocket.

On receiving this new information, the client updates the model().
Each model() update now triggers a redraw of the whole UI.

Misc:

- add helper function for filtering messages.
- add merging code for messages in Model() so we can process message edits.
- borrow and modify queues implementation from zulip-js.