Add goodbye button.
This is mostly intended to prototype a concept
of modal screens.

But it may also be the way that we let folks
have control over reading messages.  (For
unreads, I think we might want to err on the
side of leaving messages read until somebody
explicitly leaves the conversation.  This
reduces a bit of traffic and gives the user
more control.  And at least for PMs, there
really isn't much unread-count spam, unless
you're having a bunch of different PM convos.)