css: Add padding/border for layout panes.
This is less than ideal, but we need some kind of
padding and borders between the layout panes.

I'm also not crazy about centering the single
pane (if there is only one).  I would rather
left-justify it and then put a big button in
the second pane that says "Add a second pane!!".
We can discuss, obviously.