tictactoe: Support dynamic games.
We now have a start_game message, and game keys
include user_ids.  Every time a new user joins
for the first time, they create a new game.  (We
handle the new-game creation server-side.)

This is still pretty limited--we want users to
be able to "close" games and then start new ones
from the UI.