Eliminate 5-second delay for queue_id.
We also store queue_id in public/events now,
since it's not really part of the Zulip model,
and it's more of a transient thing related
to the session.

The 5-second delay was unnecessary, because
we don't even get events from the server
until the client has established a websocket
connection to us.

The actual problem here was that we were
needlessly waiting too long to set up listeners.

Note that we now wait until the websocket
connection is open before we start the UI.

Other minor things touched here:

    - we don't need page_params.ws_port
    - send_model_update -> send_to_client