layout: Color the panes for better tracking.
When the panes are all white and the layout changes move them around,
it is hard to keep track of which pane went where. This way, there is
some additional visual cue to help the user out.

We should have a better strategy than this one, mahybe not coloring
the whole pane and just setting a border. Let's experiment and see
what we like best.