layout: Implement pane titles.
We could have implemented them by making the update and render
functions of each widget bubble up the titles all the way to
the top level layout, or...

We use a setTimeout to wait for elements to get added to the DOM
and then set the title. All we need for this is one jquery DOM
element to do this. The helper function abstracts away all this
detail from the callers, and also allows the title to be updated
on the fly.