authentication: Support the use of api keys.
This allows us to connect to arbitrary Zulip
realms with an email and api key, which means
we don't need the server to support oauth
for us to use the proxy.

The flow is that if you attempt to load the
app without an active session, we send you
to a login page where you enter your email
(really "delivery_email") and api_key and
hit "Sign in".  And then you're redirected to
the home page.  We store your info on the session
until you log out.

We don't yet gracefully handle the case that
your credentials are wrong.

Note that this still allows multiple users on
the same instance of the proxy, so users can
interact outside of Zulip, such as playing games.

The api_key approach is still inferior to oauth
from a security standpoint. We don't have to keep
api_keys for all our users in any kind of permanent
configuration files (good), but we do have
them in session memory (bad). So we'll eventually want
to add some lingo that you must trust the person
running the proxy and authorize them to use your
api_key on their behalf.

Also, from a convenience standpoint, it is a
pain for users to have to go get their api key
instead of just using the oauth flow.

It's possible that we want the new boolean flag
here to be something like `use_oauth` instead
of `use_api_key`.  Also, we may decide to allow
both login methods some day, although I don't
actually see why you'd ever want both options.

We might want to get cute with some kind
of `credentials` encapsulation that holds either
the email/api_key or the oauth access token.
For now the code reads a little sloppy, but it's
still just two cases, and there are only a few
places in the code where we have to handle both
possibilities.

Here is my config for running locally against
czo (which doesn't have oauth support yet):

module.exports = {
    host: 'http://showell.zulipdev.org',
    port: 3000,
    app_url: 'https://chat.zulip.org',
    use_api_key: true,
    session_secret: "somesecret",
};

Finally, we want to make the login form nicer
both visually and code-wise (i.e. move it to a
pug).