Prevent errors from missing cross-realm bots.
This adds a generic by_id_maybe helper
for all our models objects built on
ZulipAccessor.

And we use this for finding users.

When we are searching for users, it is possible
that some users who are senders in messages may
not be among the users we fetched, and we can
usually tolerate this.

The current use case is that the notification bot
is not sent in the /users endpoint. The docs don't
specifically call this out, but I believe that you
only get bots from your realm, and cross-realm
bots technically aren't in your realm. (As an
aside, I have wanted notification bots to just
be in your own realm since forever ago.)

We might also hit future minor edge cases with
deleted users or other special kind of guest-like
users.