games: Remove tictactoe support for now.
The only technically interesting piece that was
driven out by tictactoe was websockets, and we
now use websockets for Zulip data, albeit only
in one direction.

Tictactoe is kinda neat in that it does peer to
peer stuff from the client's perspective and
two-way messaging from the server's perspective.

We will try to resurrect this some day.

This diff (in reverse) is also interesting to
see some key integration points in the app.