Require explicit click in split_pane widget.
This makes it less noisy for the user (for example,
I don't want to see an "announce" message list),
plus it makes it so that on page load we defer
some work (which concerns me more for stuff like console
noise than actual performance).

Note that the split_pane widgets are still sticky
from the perspective of the tab_bar.  We may
eventually want to just have the tab_bar be more
like an app launcher and when you move away from
a tab, we effectively close it.  (This is tied
to some of our upcoming re-think on how the layout
widget works.)