Add local echo for stream messages.
Thanks to all the prep commits, this is mostly
a matter of extracting filter_messages and
having it work for both pms and streams.

Then we add local_echo for the transmit piece,
and we call in_flight_messages from the topic
view code.

Note my comment that we might want to move
filter_messages out of transmit.js, since
it is generic in nature and could help us
build the actual message views.

The logic in find_pms_with, find_topics_for,
etc. is very similar, with the annoyance
that we use "topic" outbound but check
"subject" inbound.